+++ src/main/java/com/xiaoleilu/hulu/view/extra/VelocityView.java
@@ -1,11 +1,12 @@
-package com.xiaoleilu.hulu.view;
+package com.xiaoleilu.hulu.view.extra;
 
 import javax.servlet.http.HttpServletResponse;
 
 import com.xiaoleilu.hulu.Request;
 import com.xiaoleilu.hulu.Response;
+import com.xiaoleilu.hulu.view.View;
+import com.xiaoleilu.hutool.extra.VelocityUtil;
 import com.xiaoleilu.hutool.util.StrUtil;
-import com.xiaoleilu.hutool.util.VelocityUtil;
 
 /**
  * Velocity内容<br>